Deduplicate some client functions
A few of these functions were pretty much a copy-paste of each other, so
I've de-duplicated these and just re-use the same block for all of them.

This improves readability a bit, and makes it a bit more obvious that all
of these actually do the same.

I've also deleted an initialiser that did nothing, since that mostly adds
noise/complexity.